+++ openstack/autoscaling/v1/instances/list.go
@@ -2,7 +2,7 @@ package instances
 
 import (
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
 )
 
 type ListOptsBuilder interface {
@@ -14,24 +14,30 @@ type ListOpts struct {
 	HealthStatus    string `q:"health_status"`
 }
 
-func (opts ListOpts) ToInstancesListQuery() (string, error) {
+func List(client *golangsdk.ServiceClient, groupID string, opts ListOpts) ([]Instance, error) {
 	q, err := golangsdk.BuildQueryString(opts)
 	if err != nil {
-		return "", err
+		return nil, err
 	}
-	return q.String(), err
+
+	raw, err := client.Get(client.ServiceURL("scaling_group_instance", groupID, "list")+q.String(), nil, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	var res []Instance
+	err = extract.IntoSlicePtr(raw.Body, &res, "scaling_group_instances")
+	return res, err
 }
 
-func List(client *golangsdk.ServiceClient, groupID string, opts ListOptsBuilder) pagination.Pager {
-	url := client.ServiceURL("scaling_group_instance", groupID, "list")
-	if opts != nil {
-		q, err := opts.ToInstancesListQuery()
-		if err != nil {
-			return pagination.Pager{Err: err}
-		}
-		url += q
-	}
-	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
-		return InstancePage{pagination.SinglePageBase(r)}
-	})
+type Instance struct {
+	ID                string `json:"instance_id"`
+	Name              string `json:"instance_name"`
+	GroupID           string `json:"scaling_group_id"`
+	GroupName         string `json:"scaling_group_name"`
+	LifeCycleStatus   string `json:"life_cycle_state"`
+	HealthStatus      string `json:"health_status"`
+	ConfigurationName string `json:"scaling_configuration_name"`
+	ConfigurationID   string `json:"scaling_configuration_id"`
+	CreateTime        string `json:"create_time"`
 }
